fix(fault-proof): correct maxChallengeDuration return type to uint64 (#891)
The Solidity source returns Duration (= uint64) for both
maxChallengeDuration() and maxProveDuration(), but the hand-written
sol! binding declared uint256 for the former. The wire ABI was
unaffected (32-byte word either way), and the call site narrowed
the result back to u64 with .to::<u64>(), so no production behavior
changes — but the type-level binding now matches the contract.

Reported by Kien Nguyen.

PR #809 supersedes this fix by replacing all hand-written sol!
bindings with forge-generated ones, eliminating this class of drift.

+++ fault-proof/src/contract.rs
@@ -97,7 +97,7 @@ sol! {
         function gameOver() external view returns (bool gameOver_);
 
         /// @notice Returns the max challenge duration.
-        function maxChallengeDuration() external view returns (uint256 maxChallengeDuration_);
+        function maxChallengeDuration() external view returns (uint64 maxChallengeDuration_);
 
         /// @notice Returns the max prove duration.
         function maxProveDuration() external view returns (uint64 maxProveDuration_);
